% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_size.R
\name{general_sir}
\alias{general_sir}
\title{General SIR ode}
\usage{
general_sir(t, y, pars, C, Nage, Ntiter)
}
\arguments{
\item{t}{current time, as for ode function}

\item{pars}{vector of R0, 1/gamma and alphas (vector of immunity conferred for each titre class)}

\item{C}{the normalised contact matrix of contact frequencies between each age and titre classes}

\item{Nage}{number of age classes}

\item{Ntiter}{number of titre classes}

\item{vector}{of compartment sizes eg. S11, I11, R11, S12, I12, R12 etc...}
}
\value{
a list of compartment size changes, as required by deSolve
}
\description{
Generic SIR ode function (for use in deSolve), taking an arbritrary number of populations, ages and immunity classes
}

