% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_size.R
\name{setup_C}
\alias{setup_C}
\title{Normalised contact rate matrix}
\usage{
setup_C(C1, Ns)
}
\arguments{
\item{C1}{the non-normalised contact matrix of contact frequencies between each age class}

\item{Ns}{the matrix of population sizes for each age/titre combination (non-normalised)}
}
\value{
a matrix of the normalised contact rates between all age/titre class combinations.
}
\description{
Given a matrix of contact frequency between age classes and a matrix of true population sizes, gives the scaled, normalised contact rate matrix for each age/titre class.
}

