% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_size.R
\name{sir}
\alias{sir}
\title{Single population SIR case - ODE model}
\usage{
sir(t, y, pars, C)
}
\arguments{
\item{t}{current time, as for ode function}

\item{pars}{vector of R0 and duration of infectious period}

\item{C}{the contact rate}

\item{vector}{of compartment sizes ie. S, I and R}
}
\value{
a list of compartment size changes, as required by deSolve
}
\description{
1 age class SIR ode function (for use in deSolve)
}
\seealso{
\code{\link{sir_3}}, \code{\link{sir_2}}
}

