% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_size.R
\name{sir_2}
\alias{sir_2}
\title{Two population SIR case - ODE model}
\usage{
sir_2(t, y, pars, C)
}
\arguments{
\item{t}{current time, as for ode function}

\item{pars}{vector of R0, alpha (immunity) and duration of infectious period}

\item{C}{the 2x2 contact rate matrix}

\item{vector}{of compartment sizes ie. S1, I1, R1, S2 etc...}
}
\value{
a list of compartment size changes, as required by deSolve
}
\description{
2 age class SIR ode function (for use in deSolve)
}
\seealso{
\code{\link{sir_3}}, \code{\link{sir}}
}

