% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_size.R
\name{sir_22}
\alias{sir_22}
\title{Two age/two titre class SIR ODE function}
\usage{
sir_22(t, y, pars, C)
}
\arguments{
\item{t}{current time, as for ode function}

\item{pars}{vector of R0, alpha1, alpha2 (immunity of titre class 1 and 2) and duration of infectious period}

\item{C}{the contact rate matrix for contact rates between each age/titre class (4x4 matrix)}

\item{vector}{of compartment sizes ie. S11, I11, R11, S12 etc...}
}
\value{
a list of compartment size changes, as required by deSolve
}
\description{
Calculates the final size of an epidemic given 2-dimensional population categorisation eg. age and immunity class using an SIR model. Only takes 2 ages and 2 titres.
}
\seealso{
\link{\code{epi_ode_size}}
}

