% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_size.R
\name{sir_3}
\alias{sir_3}
\title{3 age class SIR ode}
\usage{
sir_3(t, y, pars, C)
}
\arguments{
\item{t}{current time, as for ode function}

\item{pars}{vector of R0, alpha (immunity) and duration of infectious period}

\item{C}{the normalised contact matrix of contact frequencies between each age classes}

\item{vector}{of compartment sizes eg. S1, I1, R1, S2, I2, R2 etc...}
}
\value{
a list of compartment size changes, as required by deSolve
}
\description{
3 age class SIR ode function (for use in deSolve)
}
\seealso{
\code{\link{sir}}, \code{\link{sir_2}}
}

